+++ test/hexapdf/layout/test_container_box.rb
@@ -10,6 +10,10 @@ describe HexaPDF::Layout::ContainerBox do
     @frame = HexaPDF::Layout::Frame.new(0, 0, 100, 100)
   end
 
+  def child_box(height: 0, width: 0, **style_properties)
+    @doc.layout.box(height: height, width: width, style: style_properties)
+  end
+
   def create_box(children, **kwargs)
     HexaPDF::Layout::ContainerBox.new(children: Array(children), **kwargs)
   end
@@ -48,16 +52,40 @@ describe HexaPDF::Layout::ContainerBox do
 
   describe "fit_content" do
     it "fits the children according to their mask_mode, valign, and align style properties" do
-      box = create_box([@doc.layout.box(height: 20),
-                        @doc.layout.box(height: 20, style: {valign: :bottom, mask_mode: :fill_horizontal}),
-                        @doc.layout.box(width: 20, style: {align: :right, mask_mode: :fill_vertical})])
+      box = create_box([child_box(height: 20),
+                        child_box(height: 20, valign: :bottom, mask_mode: :fill_horizontal),
+                        child_box(width: 20, align: :right, mask_mode: :fill_vertical)])
       check_box(box, 100, 100, [[0, 80], [0, 0], [80, 20]])
     end
 
     it "respects the initially set width/height as well as border/padding" do
-      box = create_box(@doc.layout.box(height: 20), height: 50, width: 40,
+      box = create_box(child_box(height: 20), height: 50, width: 40,
                        style: {padding: 2, border: {width: 3}})
       check_box(box, 40, 50, [[5, 75]])
+    end
+
+    it "fails if splitting is not allowed and the content is too big" do
+      box = create_box([child_box(height: 80), child_box(height: 30)])
+      box.fit(@frame.available_width, @frame.available_height, @frame)
+      assert(box.fit_result.failure?)
+    end
+
+    it "splits the box if splitting is allowed and the content is too big" do
+      box = create_box([child_box(height: 80), child_box(height: 30)], splitable: true)
+      box.fit(@frame.available_width, @frame.available_height, @frame)
+      assert(box.fit_result.overflow?)
+    end
+  end
+
+  describe "split_content" do
+    it "assigns the overflown boxes to the split box" do
+      box = create_box([child_box(height: 80), child_box(height: 30)], splitable: true)
+      box.fit(@frame.available_width, @frame.available_height, @frame)
+      assert(box.fit_result.overflow?)
+      box_a, box_b = box.split
+      assert_same(box, box_a)
+      assert(box_b.split_box?)
+      assert_equal(1, box_b.children.size)
     end
   end
 
@@ -67,10 +95,10 @@ describe HexaPDF::Layout::ContainerBox do
     end
 
     it "draws the result onto the canvas" do
-      child_box = @doc.layout.box(height: 20) do |canvas, b|
+      cbox = @doc.layout.box(height: 20) do |canvas, b|
         canvas.line(0, 0, b.content_width, b.content_height).end_path
       end
-      box = create_box(child_box)
+      box = create_box(cbox)
       box.fit(100, 100, @frame)
       box.draw(@canvas, 0, 50)
       assert_operators(@canvas.contents, [[:save_graphics_state],